+++ lib/amazing_activist/broken_contract_error.rb
@@ -1,7 +1,5 @@
 # frozen_string_literal: true
 
-require_relative "./error"
-
 module AmazingActivist
   class BrokenContractError < Error; end
 end