+++ lib/amazing_activist/unwrap_error.rb
@@ -1,7 +1,5 @@
 # frozen_string_literal: true
 
-require_relative "./error"
-
 module AmazingActivist
   class UnwrapError < Error
     # @return [Outcome::Failure]