+++ spec/spec_helper.rb
@@ -1,5 +1,10 @@
 # frozen_string_literal: true
 
+require "amazing-activist"
+
+# Ensure project struture and Zeitwerk ara correct
+AmazingActivist::Loader.eager_load(force: true)
+
 require "support/amazing_activist"
 
 # See https://rubydoc.info/gems/rspec-core/RSpec/Core/Configuration