+++ CHANGES.md
@@ -7,12 +7,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+
+## [0.9.0] - 2025-05-28
+
 ### Changed
 
 - (BREAKING) Switch to Zeitwerk autoloader.
 - (BREAKING) Deprecate `AmazingActivist::Outcome::{Success,Failure}`
   in favour of `AmazingActivist::Success` and `AmazingActivist::Failure`
   respectively.
+- (BREAKING) Harden outcomes initializer API
+
 
 ## [0.8.0] - 2025-05-01
 
@@ -115,7 +120,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial release.
 
-[Unreleased]: https://github.com/ixti/amazing-activist/compare/v0.8.0...main
+[Unreleased]: https://github.com/ixti/amazing-activist/compare/v0.9.0...main
+[0.9.0]: https://github.com/ixti/amazing-activist/compare/v0.8.0...v0.9.0
 [0.8.0]: https://github.com/ixti/amazing-activist/compare/v0.7.0...v0.8.0
 [0.7.0]: https://github.com/ixti/amazing-activist/compare/v0.6.0...v0.7.0
 [0.6.0]: https://github.com/ixti/amazing-activist/compare/v0.5.2...v0.6.0